requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-text-splitters==0.0.1
chromadb==0.4.22
pypdf==3.17.0
# python-dotenv==1.0.0
openai==1.3.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
TA-Lib>=0.4.24
pyyaml>=6.0
python-dotenv>=1.0.0
